Refactor test suite to remove "magic"
This refactor removes the test magic for "merge" operations.
I attempted to make the test code as concise as possible since it is pure boilerplate with this method.
It does, however, give us the critical ability to address corner cases and add tests that need to rase exceptions, for example.
A recent [blog post](https://testing.googleblog.com/2019/12/testing-on-toilet-tests-too-dry-make.html) backs up this line of thought, though I don't want to be accused of "monkey see, monkey do".  :)